PAD-Lab3/rest/requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar
django-crispy-forms
ipdb
django-extensions
flake8
mkdocs
